+++ src/ec-evaluator/utils.ts
@@ -6,7 +6,7 @@ import * as errors from '../errors/errors'
 import { RuntimeSourceError } from '../errors/runtimeSourceError'
 import Closure from '../interpreter/closure'
 import { Environment, Frame, Value } from '../types'
-import * as ast from '../utils/astCreator'
+import * as ast from '../utils/ast/astCreator'
 import * as instr from './instrCreator'
 import { Agenda } from './interpreter'
 import { AgendaItem, AppInstr, AssmtInstr, Instr, InstrType } from './types'
