+++ src/gpu/transfomer.ts
@@ -1,6 +1,6 @@
 import * as es from 'estree'
 
-import * as create from '../utils/astCreator'
+import * as create from '../utils/ast/astCreator'
 import { ancestor, make, simple } from '../utils/walkers'
 import GPUBodyVerifier from './verification/bodyVerifier'
 import GPULoopVerifier from './verification/loopVerifier'
