+++ src/infiniteLoops/instrument.ts
@@ -2,7 +2,7 @@ import { generate } from 'astring'
 import * as es from 'estree'
 
 import { transformImportDeclarations } from '../transpiler/transpiler'
-import * as create from '../utils/astCreator'
+import * as create from '../utils/ast/astCreator'
 import { recursive, simple, WalkerCallback } from '../utils/walkers'
 // transforms AST of program
