+++ src/validator/__tests__/validator.ts
@@ -3,7 +3,7 @@ import * as es from 'estree'
 import { mockContext } from '../../mocks/context'
 import { parse } from '../../parser/parser'
 import { Chapter, NodeWithInferredType } from '../../types'
-import { getVariableDeclarationName } from '../../utils/astCreator'
+import { getVariableDeclarationName } from '../../utils/ast/astCreator'
 import { stripIndent } from '../../utils/formatters'
 import { expectParsedError } from '../../utils/testing'
 import { simple } from '../../utils/walkers'
